refactor(thor): split validate_wave tests under 300-line cap
CI file-size guard caught tests/validate.rs at 392 lines (cap
300). Same pattern previously used for pr.rs / pr_sync.rs in PR
T2.04: extract the new test surface to a sibling file.

## Problem

Previous T3.06 commit added 4 wave-scoped tests + 2 helpers to
the existing tests/validate.rs, pushing it from 248 to 392
lines. CI fmt+clippy+test failed on the file-size guard step
(it does enforce the cap on tests/, not just src/, contrary to
the convention assumed in the prior commit narrative).

## What changed

- New tests/validate_wave.rs (162 lines) holds the 4 T3.06
  tests + the `add_task_in_wave` and `submit_with_evidence`
  helpers. Re-imports the same fresh() helper pattern locally.
- tests/validate.rs back to 249 lines (unchanged content).

No production code touched. No test removed. No semantics
changed.

## Validation

- `cargo fmt --all -- --check` clean.
- `cargo clippy --workspace --all-targets -D warnings` clean.
- `cargo test -p convergio-thor`: 9 (validate.rs) + 4
  (validate_wave.rs) = **13 passed, 0 failed**.
- Both test files now under the 300-line cap.

## Files touched

- `crates/convergio-thor/tests/validate.rs` (392 → 249, removed split content)
- `crates/convergio-thor/tests/validate_wave.rs` (new, 162)

Tracks: dd88af1d-2d42-4602-a9d3-abd7aa92ae94
